Create README.md
Database of ECGs and CAGs
Introduction
Background: Electrical storm (ES) is a heterogeneous clinical emergency that can present with malignant ventricular arrhythmias such as ventricular fibrillation (VF),
ventricular tachycardia (VT), requiring the need for cardiac defibrillation. ES is a life-threatening condition with a high mortality rate. Successfully managing ES
in the setting of acute myocardial infarction（MI）is expected to be known by physicians on call to reduce in-hospital mortality.

Case presentation: A 57-year-old man presenting with acute onset chest pain was found to have an infero-posterior ST-segment elevation myocardial infarction (STEMI) 
complicated by acute right ventricular MI secondary to total occlusion of the proximal right coronary artery (RCA). The patient developed ES in the form of recurrent
VF that was managed successfully with electrical defibrillation, antiarrhythmic therapy with amiodarone and esmolol, endotracheal intubation, sedation, electrolyte 
replacement, volume resuscitation, comfort care, psychological intervention, and percutaneous coronary intervention (PCI) of the occluded epicardial artery. With 
these interventions used in quick succession and with the aspiration of a massive RCA thrombus, the patient was reversed to hemodynamic stability, did not have 
further episodes of VF, and survived the index hospitalization.

Conclusion: ES is a rare but fatal complication of acute MI. Residents on night shifts should be better prepared and equipped to deal with this rare condition. We 
hope our successful experience can benefit physicians on call who take care of acute MI patients that deteriorate with ES.
Data Sharing
This is the data repository of the case report sharing the ECGs and CAGs. Browse and download the data by clicking the folder links above.